Create dedicated `cli` package for `cli` classes/packages.
