+++ phenopacket-tools-cli/src/main/java/module-info.java
@@ -8,14 +8,11 @@ module org.phenopackets.phenopackettools.cli {
     requires org.monarchinitiative.phenol.core;
     requires org.monarchinitiative.phenol.io;
 
-    requires com.google.protobuf.util;
-    requires com.fasterxml.jackson.databind;
-    requires com.fasterxml.jackson.dataformat.yaml;
-    requires commons.csv;
     requires info.picocli;
+    requires commons.csv;
     requires org.slf4j;
     requires logback.classic;
 
-    opens org.phenopackets.phenopackettools.command to info.picocli;
-    opens org.phenopackets.phenopackettools.command.validate to info.picocli;
+    opens org.phenopackets.phenopackettools.cli.command to info.picocli;
+    opens org.phenopackets.phenopackettools.cli.command.validate to info.picocli;
 }
