+++ phenopacket-tools-cli/src/main/java/org/phenopackets/phenopackettools/cli/examples/PhenopacketExample.java
@@ -1,0 +1,7 @@
+package org.phenopackets.phenopackettools.cli.examples;
+
+import org.phenopackets.schema.v2.Phenopacket;
+
+interface PhenopacketExample {
+    Phenopacket getPhenopacket();
+}
