--- phenopacket-tools-cli/src/main/java/org/phenopackets/phenopackettools/examples/BethlehamMyopathy.java
@@ -1,118 +1,0 @@
-package org.phenopackets.phenopackettools.examples;
-
-
-import org.ga4gh.vrsatile.v1.GeneDescriptor;
-import org.ga4gh.vrsatile.v1.VcfRecord;
-import org.phenopackets.phenopackettools.builder.PhenopacketBuilder;
-import org.phenopackets.phenopackettools.builder.builders.*;
-import org.phenopackets.phenopackettools.builder.constants.Status;
-import org.phenopackets.schema.v2.Phenopacket;
-import org.phenopackets.schema.v2.core.GenomicInterpretation;
-
-import static org.phenopackets.phenopackettools.builder.builders.OntologyClassBuilder.ontologyClass;
-
-public class BethlehamMyopathy implements PhenopacketExample {
-    private static final String PHENOPACKET_ID = "arbitrary proband id";
-    private static final String INTERPRETATION_ID = "arbitrary interpretation id";
-    private static final String PROBAND_ID = "proband A";
-
-    private final Phenopacket phenopacket;
-
-    public BethlehamMyopathy() {
-        var authorAssertion = EvidenceBuilder.authorStatementEvidence("PMID:30808312", "COL6A1 mutation leading to Bethlem myopathy with recurrent hematuria: a case report");
-        var bethlehamMyopathy = ontologyClass("OMIM:158810", "Bethlem myopathy 1");
-        var individual = IndividualBuilder.builder(PROBAND_ID).male().ageAtLastEncounter("P6Y3M").build();
-        var metaData = MetaDataBuilder.builder("2021-05-14T10:35:00Z", "anonymous biocurator")
-                .addResource(Resources.hpoVersion("2021-08-02"))
-                .addResource(Resources.genoVersion("2020-03-08"))
-                .addExternalReference(authorAssertion.getReference())
-                .build();
-
-        var genomicInterpretation = COL6A1variant();
-
-        var diagnosis = DiagnosisBuilder.builder(bethlehamMyopathy).addGenomicInterpretation(genomicInterpretation).build();
-        var interpretation = InterpretationBuilder.builder(INTERPRETATION_ID).completed(diagnosis);
-        var ventricularSeptalDefect =
-                PhenotypicFeatureBuilder.builder("HP:0001629", "Ventricular septal defect")
-                        .congenitalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var coarseFacial =
-                PhenotypicFeatureBuilder.builder("HP:0000280", "Coarse facial features")
-                        .addEvidence(authorAssertion)
-                        .build();
-        var cryptorchidism =
-                PhenotypicFeatureBuilder.builder("HP:0008689", "Bilateral cryptorchidism")
-                        .congenitalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var polyhydramnios =
-                PhenotypicFeatureBuilder.builder("HP:0001561", "Polyhydramnios")
-                        .fetalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var micropenis =
-                PhenotypicFeatureBuilder.builder("HP:0000054", "Micropenis")
-                        .congenitalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var anonychia =
-                PhenotypicFeatureBuilder.builder("HP:0001798", "Anonychia")
-                        .congenitalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var vermisHypoplasia =
-                PhenotypicFeatureBuilder.builder("HP:0001320", "Cerebellar vermis hypoplasia")
-                        .addEvidence(authorAssertion)
-                        .build();
-        var cataract =
-                PhenotypicFeatureBuilder.builder("HP:0000518", "Cataract")
-                        .infantileOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        var dilatedFourthVentricle =
-                PhenotypicFeatureBuilder.builder("HP:0002198", "Dilated fourth ventricle")
-                        .addEvidence(authorAssertion)
-                        .build();
-        var unilateralCleftLip =
-                PhenotypicFeatureBuilder.builder("HP:0100333", "Unilateral cleft lip")
-                        .congenitalOnset()
-                        .addEvidence(authorAssertion)
-                        .build();
-        phenopacket = PhenopacketBuilder.create(PHENOPACKET_ID, metaData)
-                .individual(individual)
-                .addPhenotypicFeature(ventricularSeptalDefect)
-                .addPhenotypicFeature(coarseFacial)
-                .addPhenotypicFeature(cryptorchidism)
-                .addPhenotypicFeature(polyhydramnios)
-                .addPhenotypicFeature(micropenis)
-                .addPhenotypicFeature(anonychia)
-                .addPhenotypicFeature(vermisHypoplasia)
-                .addPhenotypicFeature(cataract)
-                .addPhenotypicFeature(dilatedFourthVentricle)
-                .addPhenotypicFeature(unilateralCleftLip)
-                .addInterpretation(interpretation)
-                .build();
-    }
-
-
-    private GenomicInterpretation COL6A1variant() {
-        var variationDescriptor =
-                VariationDescriptorBuilder.builder("variant id")
-                        .heterozygous()
-                        .hgvs("NM_001848.2:c.877G>A")
-                        .geneContext(GeneDescriptorBuilder.of("HGNC:2211", "COL6A1"))
-                        .vcfHg38("chr21",45989626, "G","A")
-                        .build();
-        var col6a1VariantInterpretation =
-                VariantInterpretationBuilder.of(variationDescriptor, Status.pathogenic());
-        return GenomicInterpretationBuilder.builder(INTERPRETATION_ID)
-                .causative()
-                .variantInterpretation(col6a1VariantInterpretation).build();
-    }
-
-    @Override
-    public Phenopacket getPhenopacket() {
-        return phenopacket;
-    }
-}
