--- phenopacket-tools-cli/src/main/java/org/phenopackets/phenopackettools/examples/PhenopacketExample.java
@@ -1,7 +1,0 @@
-package org.phenopackets.phenopackettools.examples;
-
-import org.phenopackets.schema.v2.Phenopacket;
-
-interface PhenopacketExample {
-    Phenopacket getPhenopacket();
-}
