--- phenopacket-tools-cli/src/test/java/org/phenopackets/phenopackettools/command/BasePTCommandTest.java
@@ -1,16 +1,0 @@
-package org.phenopackets.phenopackettools.command;
-
-import org.junit.jupiter.api.Test;
-
-import static org.hamcrest.MatcherAssert.*;
-import static org.hamcrest.Matchers.*;
-
-public class BasePTCommandTest {
-
-    @Test
-    public void markIsSupportedForStdin() {
-        // We need this functionality when sniffing format from STDIN.
-        assertThat(System.in.markSupported(), equalTo(true));
-    }
-
-}
